add shiro

+++ el-rapid-generator/el-genreator-war/src/main/resources/template/${project_name}/${project_name}-api/src/main/java/${basepackage_dir}/api/${className}Service.java
@@ -27,7 +27,7 @@ public interface ${className}Service{
 	@POST
 	@Produces({MediaType.APPLICATION_JSON})
     @Consumes({MediaType.APPLICATION_JSON})
-	ResultDTO<Integer> insert${className}(${className}Dto ${classNameLower}Dto);
+	ResultDTO<Long> insert${className}(${className}Dto ${classNameLower}Dto);
 	 /**  
      *   
      * <update one>
